+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/pde_mgmt/dto/PlantDescriptionUpdate.java
@@ -19,6 +19,6 @@ public interface PlantDescriptionUpdate {
     Optional<Boolean> active();
     Optional<List<PdeSystem>> systems();
     Optional<List<Connection>> connections();
-    Optional<List<Float>> include();
+    Optional<List<Integer>> include();
 
 }
